Re-configure elements
I wanted to change which main sections of the page managed what in the
name of semantic HMTL, so I've split the navigation elements into its
own specific navigation bar, and set the header to be the main banner of
the page. This leaves the rest of the page's main content to lie within
the <main> element (Not used in the landing page).